la setstatus endpointent inni ForslagControlleren
